p3_field

Module extension

Structs

BinomialExtensionField

PackedBinomialExtensionField

Traits

BinomiallyExtendable

Trait for fields that support binomial extension of the form F[X]/(X^D - W).

BinomiallyExtendableAlgebra

Trait for algebras which support binomial extensions of the form A[X]/(X^D - W) with W in the base field F.

ComplexExtendable

A field for which p = 3 (mod 4). Equivalently, -1 is not a square, so the complex extension can be defined F[i] = F[X]/(X^2+1).

HasComplexBinomialExtension

The complex extension of this field has a binomial extension.

HasFrobenius

Trait for extension fields that support Frobenius automorphisms.

HasTwoAdicBinomialExtension

Trait for binomial extensions that support a two-adic subgroup generator.

HasTwoAdicComplexBinomialExtension

The complex extension of this field has a two-adic binomial extension.

Functions

octic_mul

Multiplication in an octic binomial extension field.

quartic_mul

Multiplication in a quartic binomial extension field.

quintic_mul

Multiplication in a quintic binomial extension field.

vector_add

Add two vectors element wise.

Type Aliases

Complex