p3_field::extension

Struct BinomialExtensionField

pub struct BinomialExtensionField<F, const D: usize, A = F> { /* private fields */ }

Implementations

impl<F, A, const D: usize> BinomialExtensionField<F, D, A>

pub const fn new(value: [A; D]) -> Self

Create an extension field element from an array of base elements.

Any array is accepted. No reduction is required since base elements are already valid field elements.

impl<F: Copy, const D: usize> BinomialExtensionField<F, D, F>

pub const fn new_array<const N: usize>(input: [[F; D]; N]) -> [Self; N]

Convert a [[F; D]; N] array to an array of extension field elements.

Const version of input.map(BinomialExtensionField::new).

Panics

Panics if N == 0.

impl<R: PrimeCharacteristicRing> BinomialExtensionField<R, 2>

Convenience methods for complex extensions

pub const fn new_complex(real: R, imag: R) -> Self

pub const fn new_real(real: R) -> Self

pub const fn new_imag(imag: R) -> Self

pub fn real(&self) -> R

pub fn imag(&self) -> R

pub fn conjugate(&self) -> Self

pub fn norm(&self) -> R

pub fn to_array(&self) -> [R; 2]

pub fn rotate<Ext: Algebra<R>>(&self, rhs: &Complex<Ext>) -> Complex<Ext>