p3_field::op_assign_macros

Macro impl_sub_base_field

macro_rules! impl_sub_base_field {
    ($alg_type:ty, $field_type:ty $(, ($type_param:ty, $param_name:ty))?) => { ... };
}

Given two structs Alg and Field where Alg implements From<Field>, implement Sub<Field> for Alg and Sub<Alg> for Field.

All are implemented in the simplest way by using From to map the Field element to an Alg element and then applying the native sub methods on Alg elements.