p3_field

Struct FieldArray

#[repr(transparent)]
pub struct FieldArray<F: Field, const N: usize>(pub [F; N]);

Tuple Fields
0: [F; N]
Implementations

impl<F: Field, const N: usize> FieldArray<F, N>

pub fn map<Func, U: Field>(self, f: Func) -> FieldArray<U, N>
where Func: FnMut(F) -> U,

Apply a function to each element, returning a new FieldArray<U, N>.

pub fn map_into_array<Func, U>(self, f: Func) -> [U; N]
where Func: FnMut(F) -> U,

Apply a function to each element, returning a raw array [U; N].

Unlike map, this does not require U: Field.

pub const fn as_raw_slice(s: &[Self]) -> &[[F; N]]

View as a slice of raw [F; N] arrays.

This is a zero-cost transmute enabled by the #[repr(transparent)] layout.

pub const fn as_raw_slice_mut(s: &mut [Self]) -> &mut [[F; N]]

View as a mutable slice of raw [F; N] arrays.

This is a zero-cost transmute enabled by the #[repr(transparent)] layout.

pub const fn from_raw_slice(s: &[[F; N]]) -> &[Self]

Reinterpret a slice of [F; N] as a slice of FieldArray<F, N>.

This is a zero-cost transmute enabled by the #[repr(transparent)] layout.

pub const fn from_raw_slice_mut(s: &mut [[F; N]]) -> &mut [Self]

Reinterpret a mutable slice of [F; N] as a mutable slice of FieldArray<F, N>.

This is a zero-cost transmute enabled by the #[repr(transparent)] layout.