p3_field

Trait Packable

pub trait Packable:
    'static
    + Default
    + Copy
    + Send
    + Sync
    + PartialEq
    + Eq { }

A trait to constrain types that can be packed into a packed value.

The Packable trait allows us to specify implementations for potentially conflicting types.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.