p3_field

Trait PermutationMonomial

pub trait PermutationMonomial<const N: u64>: InjectiveMonomial<N> {
    // Required method
    fn injective_exp_root_n(&self) -> Self;
}

A ring implements PermutationMonomial<N> if the algebraic function f(x) = x^N is invertible and thus acts as a permutation on elements of the ring.

In all cases we care about, this means that we can find another integer K such that x = x^{NK} for all elements of our ring.

Required Methods

fn injective_exp_root_n(&self) -> Self

Compute x -> x^K for a given K > 1 such that x^{NK} = x for all elements x.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.