p3_field

Module extension

Structs

BinomialExtensionField

Traits

BinomiallyExtendable

Binomial extension field trait. A extension field with a irreducible polynomial X^d-W such that the extension is F[X]/(X^d-W).

ComplexExtendable

A field for which p = 3 (mod 4). Equivalently, -1 is not a square, so the complex extension can be defined F[X]/(X^2+1).

HasComplexBinomialExtension

The complex extension of this field has a binomial extension.

HasFrobenius

HasTwoAdicBionmialExtension

Optional trait for implementing Two Adic Binomial Extension Field.

HasTwoAdicComplexBinomialExtension

The complex extension of this field has a two-adic binomial extension.

Type Aliases

Complex